Use absolute paths to include modular Xresources. Note:
worlddevelopment adopted the Void standard to create a user "anon".

It is a general account name without spaces and short, complementing
root.

This is useful to avoid redundancy, i.e. on a server world developers
share users or in case of SSL key authorization have no home directory.
Also it is less error prone in case of absolute paths (e.g. by accident).